(SERVER) Refactor folder naming and router to get route path from folder name
