+++ server/src/__tests__/integration/user.integration.test.ts
@@ -10,8 +10,8 @@ import { Model } from 'mongoose';
 import { IUser } from '../../models/user';
 import App from '../../app';
 import DIContainer from '../../di/diContainer';
-import PostRouter from '../../modules/posts/PostRouter';
-import UserRouter from '../../modules/users/UserRouter';
+import PostRouter from '../../api/posts/PostRouter';
+import UserRouter from '../../api/users/UserRouter';
 import DI_TYPES from '../../di/DITypes';
 import IDatabase from '../../database/interfaces/IDatabase';
 
@@ -29,8 +29,8 @@ describe('User Routes', () => {
   beforeAll(async () => {
 
     app = await new App().config([
-      new PostRouter('posts'),
-      new UserRouter('users')
+      new PostRouter(),
+      new UserRouter()
     ]);
     const database = DIContainer.get<IDatabase>(DI_TYPES.Database)
