+++ server/src/api/posts/interfaces/IPostController.ts
