+++ server/src/api/posts/interfaces/IPostService.ts
