+++ server/src/api/posts/post.schema.ts
